requirements.txt
Django==4.1.7
# django-phonenumber-field==7.0.2
djangorestframework==3.14
